perf: add populate options
Before this, running getComicBy[slug|id] caused to return all images of
all issues. This caused really big performance issue in frontend.
Because from all of those images only the first image were used (for
cover image of issue) and others weren't. After this patch, we can
specify that we only want to first (or Nth) image of the issue for cover
and maybe other purposes.