feat: update image paths in blog and static pages
Adding images to blog and static pages is going to happen a lot. So to
make our job easier when writing blog posts and static pages we move
all the images to `/public/images/{type}/{name}` directory. This allows
us to serve images from different directory.

These folders also should be inside `.gitignore` because we already
have those images in `_blog/{type}/{name}/images` folders.

While doing these changes we had to modify the image's URL. To do that
we created a Remark plugin `remark-next-image`. This basically adds a
prefix to image's name.