ci: add netlify.toml file
When a page uses 'revalidate' option in 'getStaticProps', Netlify
puts that page inside a serverless function and these serverless
functions doesn't have access to filesystem.

This commit changes that by adding 'index.mdx' and 'cover_image.webp'
files in '_blog' directory to serverless function's filesystem. This is
necessary because in index page, we are requesting guides and since
that serverless function doesn't have access to '_blog' directory, it
throws an filesystem error.

TODO: We should probably remove this once @netlify/next-plugin supports
on-demand revalidation since on-demand revalidation allows us to remove
'revalidate' property and makes it static page again, hopefully.