fix: resize images based on their usage
In commit d0e2486, we disabled Vercel's automatic image optimization.
This causes images to be so large that they take up a lot of network
bandwidth. This commit uses Blogspot's image resizing feature to
resize images based on their usage.

For example, thumbnails in homepage, search page are now resized to
smaller version of original image. This reduces the image size while
loading the page.